embedded all common (old) colors from new config structure into UI added minor checks and refactoring to upper-level sdk class and activity classes
