+++ VCheckSDK/VCheckSDK/domain/VCheckDesignConfig.swift
@@ -15,7 +15,8 @@ public struct VCheckDesignConfig : Codable {
             return value
         } else {
             print("VCheckSDK - warning: Non-valid JSON was passed while initializaing VCheckDesignConfig instance; trying to set VCheck default theme...")
-            return try! JSONDecoder().decode(VCheckDesignConfig.self, from: defaultThemeJsonStr.data(using: .utf8)!)
+            return try! JSONDecoder().decode(VCheckDesignConfig.self,
+                                             from: VCheckSDKConstants.vcheckDefaultThemeConfig.data(using: .utf8)!)
         }
     }
     
@@ -33,6 +34,7 @@ public struct VCheckDesignConfig : Codable {
     let neutralHover : String?
     let neutralActive : String?
     let neutralContent : String?
+    
     let success : String?
     let successHover : String?
     let successActive : String?
@@ -48,14 +50,15 @@ public struct VCheckDesignConfig : Codable {
     let warningActive : String?
     let warningBg : String?
     let warningContent : String?
-    let base : String?
-    let base_100 : String?
-    let base_200 : String?
-    let base_300 : String?
-    let base_400 : String?
+    
+    let backgroundPrimaryColorHex : String?
+    let backgroundSecondaryColorHex : String?
+    let backgroundTertiaryColorHex : String?
+    let buttonBorderColorHex : String?
+    let sectionBorderColorHex : String?
     let base_500 : String?
-    let baseContent : String?
-    let baseSecondaryContent : String?
+    let primaryTextColorHex : String?
+    let secondaryTextColorHex : String?
     let disabled : String?
     let disabledContent : String?
 
@@ -75,6 +78,7 @@ public struct VCheckDesignConfig : Codable {
         case neutralHover = "neutralHover"
         case neutralActive = "neutralActive"
         case neutralContent = "neutralContent"
+        
         case success = "success"
         case successHover = "successHover"
         case successActive = "successActive"
@@ -90,14 +94,15 @@ public struct VCheckDesignConfig : Codable {
         case warningActive = "warningActive"
         case warningBg = "warningBg"
         case warningContent = "warningContent"
-        case base = "base"
-        case base_100 = "base_100"
-        case base_200 = "base_200"
-        case base_300 = "base_300"
-        case base_400 = "base_400"
+        
+        case backgroundPrimaryColorHex = "base"
+        case backgroundSecondaryColorHex = "base_100"
+        case backgroundTertiaryColorHex = "base_200"
+        case buttonBorderColorHex = "base_300"
+        case sectionBorderColorHex = "base_400"
         case base_500 = "base_500"
-        case baseContent = "baseContent"
-        case baseSecondaryContent = "baseSecondaryContent"
+        case primaryTextColorHex = "baseContent"
+        case secondaryTextColorHex = "baseSecondaryContent"
         case disabled = "disabled"
         case disabledContent = "disabledContent"
     }
@@ -118,6 +123,7 @@ public struct VCheckDesignConfig : Codable {
         neutralHover = try values.decodeIfPresent(String.self, forKey: .neutralHover)
         neutralActive = try values.decodeIfPresent(String.self, forKey: .neutralActive)
         neutralContent = try values.decodeIfPresent(String.self, forKey: .neutralContent)
+        
         success = try values.decodeIfPresent(String.self, forKey: .success)
         successHover = try values.decodeIfPresent(String.self, forKey: .successHover)
         successActive = try values.decodeIfPresent(String.self, forKey: .successActive)
@@ -133,62 +139,17 @@ public struct VCheckDesignConfig : Codable {
         warningActive = try values.decodeIfPresent(String.self, forKey: .warningActive)
         warningBg = try values.decodeIfPresent(String.self, forKey: .warningBg)
         warningContent = try values.decodeIfPresent(String.self, forKey: .warningContent)
-        base = try values.decodeIfPresent(String.self, forKey: .base)
-        base_100 = try values.decodeIfPresent(String.self, forKey: .base_100)
-        base_200 = try values.decodeIfPresent(String.self, forKey: .base_200)
-        base_300 = try values.decodeIfPresent(String.self, forKey: .base_300)
-        base_400 = try values.decodeIfPresent(String.self, forKey: .base_400)
+        
+        backgroundPrimaryColorHex = try values.decodeIfPresent(String.self, forKey: .backgroundPrimaryColorHex)
+        backgroundSecondaryColorHex = try values.decodeIfPresent(String.self, forKey: .backgroundSecondaryColorHex)
+        backgroundTertiaryColorHex = try values.decodeIfPresent(String.self, forKey: .backgroundTertiaryColorHex)
+        buttonBorderColorHex = try values.decodeIfPresent(String.self, forKey: .buttonBorderColorHex)
+        //TODO: review where section color hex should be replaced by button color hex!
+        sectionBorderColorHex = try values.decodeIfPresent(String.self, forKey: .sectionBorderColorHex)
         base_500 = try values.decodeIfPresent(String.self, forKey: .base_500)
-        baseContent = try values.decodeIfPresent(String.self, forKey: .baseContent)
-        baseSecondaryContent = try values.decodeIfPresent(String.self, forKey: .baseSecondaryContent)
+        primaryTextColorHex = try values.decodeIfPresent(String.self, forKey: .primaryTextColorHex)
+        secondaryTextColorHex = try values.decodeIfPresent(String.self, forKey: .secondaryTextColorHex)
         disabled = try values.decodeIfPresent(String.self, forKey: .disabled)
         disabledContent = try values.decodeIfPresent(String.self, forKey: .disabledContent)
     }
-    
-    static let defaultThemeJsonStr: String =
-     """
-     {
-       "primary": "#2E75FF",
-       "primaryHover": "#2E96FF",
-       "primaryActive": "#3361EC",
-       "primaryContent": "#FFFFFF",
-       "primaryBg": "#5D6884",
-       "accent": "#6096FF",
-       "accentHover": "#6ABFFF",
-       "accentActive": "#4F79F7",
-       "accentContent": "#FFFFFF",
-       "accentBg": "#32404A",
-       "neutral": "#FFFFFF",
-       "neutralHover": "rgba(255, 255, 255, 0.4)",
-       "neutralActive": "rgba(255, 255, 255, 0.1)",
-       "neutralContent": "#000000",
-
-       "success": "#6CFB93",
-       "successHover": "#C8FDD2",
-       "successActive": "#00DF53",
-       "successBg": "#3A4B3F",
-       "successContent": "#3B3B3B",
-       "error": "#F47368",
-       "errorHover": "#FF877C",
-       "errorActive": "#DE473A",
-       "errorBg": "#4B2A24",
-       "errorContent": "#3B3B3B",
-       "warning": "#FFB482",
-       "warningHover": "#FFBF94",
-       "warningActive": "#D3834E",
-       "warningBg": "#3F3229",
-       "warningContent": "#3B3B3B",
-
-       "base": "#2A2A2A",
-       "base_100": "#3C3C3C",
-       "base_200": "#555555",
-       "base_300": "#6A6A6A",
-       "base_400": "#7F7F7F",
-       "base_500": "#949494",
-       "baseContent": "#FFFFFF",
-       "baseSecondaryContent": "#D8D8D8",
-       "disabled": "#AAAAAA",
-       "disabledContent": "#6A6A6A"
-     }
-     """
 }
