+++ VCheckSDK/VCheckSDK/presentation/check_doc_info_stage/CheckDocInfoViewController.swift
@@ -56,7 +56,7 @@ class CheckDocInfoViewController : UIViewController {
         
         self.docFieldsTableView.dataSource = self
         
-        if let bc = VCheckSDK.shared.backgroundSecondaryColorHex {
+        if let bc = VCheckSDK.shared.designConfig!.backgroundSecondaryColorHex {
             self.docFieldsTableView.setValue(bc.hexToUIColor() , forKey: "tableHeaderBackgroundColor")
             //self.tableFooterView.backgroundColor = bc.hexToUIColor()
         }
@@ -236,7 +236,7 @@ extension CheckDocInfoViewController: UITableViewDataSource {
                 
         let field: DocFieldWitOptPreFilledData = fieldsList[indexPath.row]
         
-        if let pct = (VCheckSDK.shared.primaryTextColorHex) {
+        if let pct = (VCheckSDK.shared.designConfig!.primaryTextColorHex) {
             cell.docTextField.textColor = pct.hexToUIColor()
         }
         
@@ -282,7 +282,7 @@ extension CheckDocInfoViewController: UITableViewDataSource {
         
         if (fieldName == "date_of_birth") {
             var hintColor: UIColor? = nil
-            if let pc = VCheckSDK.shared.secondaryTextColorHex {
+            if let pc = VCheckSDK.shared.designConfig!.secondaryTextColorHex {
                 hintColor = pc.hexToUIColor()
             } else {
                 hintColor = UIColor.white
@@ -296,7 +296,7 @@ extension CheckDocInfoViewController: UITableViewDataSource {
         
         if (fieldName == "expiration_date") {
             var hintColor: UIColor? = nil
-            if let pc = VCheckSDK.shared.secondaryTextColorHex {
+            if let pc = VCheckSDK.shared.designConfig!.secondaryTextColorHex {
                 hintColor = pc.hexToUIColor()
             } else {
                 hintColor = UIColor.white
