+++ VCheckSDK/VCheckSDK/presentation/common/HeaderViewController.swift
@@ -45,19 +45,19 @@ class HeaderViewContoller: UIViewController {
      
      private func changeColorsToCustomIfPresent() {
          
-         if let backgroundHex = VCheckSDK.shared.backgroundPrimaryColorHex {
+         if let backgroundHex = VCheckSDK.shared.designConfig!.backgroundPrimaryColorHex {
              BackgroundView.appearance().backgroundColor = backgroundHex.hexToUIColor()
          }
-         if let btnsHex = VCheckSDK.shared.buttonsColorHex {
+         if let btnsHex = VCheckSDK.shared.designConfig!.primary {
              UIButton.appearance(whenContainedInInstancesOf: [BackgroundView.self]).tintColor = btnsHex.hexToUIColor()
          }
-         if let backgroundSecondaryHex = VCheckSDK.shared.backgroundSecondaryColorHex {
+         if let backgroundSecondaryHex = VCheckSDK.shared.designConfig!.backgroundSecondaryColorHex {
              VCheckSDKRoundedView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).backgroundColor = backgroundSecondaryHex.hexToUIColor()
              DocInfoViewCell.appearance(whenContainedInInstancesOf: [BackgroundView.self]).backgroundColor = backgroundSecondaryHex.hexToUIColor()
              UIView.appearance(whenContainedInInstancesOf: [DocInfoViewCell.self]).backgroundColor = UIColor.clear
              CustomizableTableView.appearance(whenContainedInInstancesOf: [VCheckSDKRoundedView.self]).backgroundColor = backgroundSecondaryHex.hexToUIColor()
          }
-         if let backgroundTertiaryHex = VCheckSDK.shared.backgroundTertiaryColorHex {
+         if let backgroundTertiaryHex = VCheckSDK.shared.designConfig!.backgroundTertiaryColorHex {
              SmallRoundedView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).backgroundColor = backgroundTertiaryHex.hexToUIColor()
              SmallRoundedView.appearance(whenContainedInInstancesOf: [DocInfoViewCell.self]).backgroundColor = backgroundTertiaryHex.hexToUIColor()
              UITextField.appearance(whenContainedInInstancesOf: [BackgroundView.self, CustomizableTableView.self]).backgroundColor = backgroundTertiaryHex.hexToUIColor()
@@ -67,22 +67,22 @@ class HeaderViewContoller: UIViewController {
              AllowedCountryViewCell.appearance(whenContainedInInstancesOf: [BackgroundView.self]).backgroundColor = backgroundTertiaryHex.hexToUIColor()
              BlockedCountryViewCell.appearance(whenContainedInInstancesOf: [BackgroundView.self]).backgroundColor = backgroundTertiaryHex.hexToUIColor()
          }
-         if let borderColorHex = VCheckSDK.shared.borderColorHex {
+         if let borderColorHex = VCheckSDK.shared.designConfig!.sectionBorderColorHex {
              SmallRoundedView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).borderColor = borderColorHex.hexToUIColor()
              VCheckSDKRoundedView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).borderColor = borderColorHex.hexToUIColor()
          }
-         if let primaryTextHex = VCheckSDK.shared.primaryTextColorHex {
+         if let primaryTextHex = VCheckSDK.shared.designConfig!.primaryTextColorHex {
              PrimaryTextView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).textColor = primaryTextHex.hexToUIColor()
              UITextField.appearance(whenContainedInInstancesOf: [BackgroundView.self, CustomizableTableView.self]).textColor = primaryTextHex.hexToUIColor()
              UIImageView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).tintColor = primaryTextHex.hexToUIColor()
              UITextView.appearance(whenContainedInInstancesOf: [DocInfoViewCell.self]).textColor = primaryTextHex.hexToUIColor()
              CustomBackgroundTextView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).textColor = primaryTextHex.hexToUIColor()
          }
-         if let secondaryTextHex = VCheckSDK.shared.secondaryTextColorHex {
+         if let secondaryTextHex = VCheckSDK.shared.designConfig!.secondaryTextColorHex {
              SecondaryTextView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).textColor = secondaryTextHex.hexToUIColor()
              NonCustomizableIconView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).tintColor = secondaryTextHex.hexToUIColor()
          }
-         if let iconsColorHex = VCheckSDK.shared.iconsColorHex {
+         if let iconsColorHex = VCheckSDK.shared.designConfig!.primary {
              CustomizableIconView.appearance(whenContainedInInstancesOf: [BackgroundView.self]).tintColor = iconsColorHex.hexToUIColor()
          }
      }
